+++ src/pages/NovaConvocacaoCandidatos/components/AdicionarNovaEscolaModal.tsx
@@ -6,17 +6,17 @@ const { Text } = Typography;
 
 import { Col, Divider, Input, Row } from "antd";
 import { Controller, useForm } from "react-hook-form";
-import { CustomFormItem } from "../../styles";
+import { CustomFormItem } from "../styles";
 import { Content } from "antd/es/layout/layout";
-import type { IConvocacaoFiltros } from "../../../../services/resources/convocacao/IConvocacao";
+import type { IConvocacaoFiltros } from "../../../services/resources/convocacao/IConvocacao";
 
 
 import { useTheme } from "styled-components";
 import {
   CustomModal,
   TextBlue,
-} from "../../../../components/estilosCompartilhados/styles";
-import AdicionarEscolaTable from "../AdicionarEscolaTable/AdicionarEscolaTable";
+} from "../../../components/estilosCompartilhados/styles";
+import AdicionarEscolaTable from "./AdicionarEscolaTable";
 
 interface INewAdicionarNovaEscolaModalProps {
   isOpen: boolean;
