+++ src/pages/NovaConvocacaoCandidatos/components/VisualizarVagasModal/VisualizarVagasModal.tsx
@@ -9,10 +9,10 @@ import { Controller, useForm } from "react-hook-form";
 import { ModalCustomFormItem as CustomFormItem, CustomModal } from "../../styles";
 import { Content } from "antd/es/layout/layout";
 import type { IConvocacaoFiltros } from "../../../../services/resources/convocacao/IConvocacao";
-import UnidadeEscolarTable from "../UnidadeEscolarTable/UnidadeEscolarTable";
+import UnidadeEscolarTable from "../UnidadeEscolarTable";
 import { TextBlue } from "../../../../components/estilosCompartilhados/styles";
 import { useState } from "react";
-import AdicionarNovaEscolaModal from "../AdicionarNovaEscolaModal/AdicionarNovaEscolaModal";
+import AdicionarNovaEscolaModal from "../AdicionarNovaEscolaModal";
 
 interface INewAdicionarNovaEscolaProps {
   isOpen: boolean;
